update vote type

+++ potato/models/tmdb-models.go
@@ -13,7 +13,7 @@ type MovieSearchResult struct {
 	ReleaseDate      string  `json:"release_date"`
 	Title            string  `json:"title"`
 	Video            bool    `json:"video"`
-	VoteAverage      int     `json:"vote_average"`
+	VoteAverage      float64 `json:"vote_average"`
 	VoteCount        int     `json:"vote_count"`
 }
